+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-All Notable changes to `artisan-cloudflare` will be documented in this file
+All notable changes to `artisan-cloudflare` will be documented in this file.
+
+## 1.1.0 - 2016-11-22
+
+- All the requests to the CloudFlare API are now asynchronous and sent concurrently.
 
 ## 1.0.0 - 2016-11-19
 
-- initial release
+- Initial release